nat: split into files
